Move SidebarTextEntry methods to SidebarPanel
